+++ src/main/java/com/agilebank/util/exceptions/InvalidAmountException.java
@@ -2,12 +2,14 @@ package com.agilebank.util.exceptions;
 
 import lombok.Getter;
 
+import java.math.BigDecimal;
+
 @Getter
 public class InvalidAmountException extends RuntimeException {
 
-  private final Double amount;
+  private final BigDecimal amount;
 
-  public InvalidAmountException(Double amount) {
+  public InvalidAmountException(BigDecimal amount) {
     super("The amount of " + amount + " is invalid; please use a non-negative amount.");
     this.amount = amount;
   }
